+++ spec/hatchet/rubies_spec.rb
@@ -51,7 +51,7 @@ describe "Ruby Versions on cedar-14" do
       expect(app.output).to match("JRUBY_OPTS is:  -Xcompile.invokedynamic=false")
       expect(app.output).not_to include("OpenJDK 64-Bit Server VM warning")
 
-      `git commit -am "redeploy" --allow-empty`
+      run!('git commit -am "redeploy" --allow-empty')
       app.set_config("JRUBY_BUILD_OPTS" => "--dev")
       app.push!
       expect(app.output).to match("JRUBY_OPTS is:  --dev")
@@ -70,11 +70,27 @@ describe "Ruby Versions on cedar-14" do
   end
 
   it "should deploy jruby with the naether gem" do
-    app = Hatchet::Runner.new("jruby_naether")
+    app = Hatchet::Runner.new("jruby_naether", stack: DEFAULT_STACK)
     app.setup!
     app.deploy do |app|
       expect(app.output).to match("Installing naether")
       expect(app.output).not_to include("An error occurred while installing naether")
+    end
+  end
+end
+
+
+describe "Upgrading ruby apps" do
+  it "works when changing from default version" do
+    app = Hatchet::Runner.new("default_ruby", stack: DEFAULT_STACK)
+    app.setup!
+    app.deploy do |app|
+      run!(%Q{echo "ruby '2.5.1'" >> Gemfile})
+      run!("git add -A; git commit -m update-ruby")
+      app.push!
+      expect(app.output).to match("2.5.1")
+      expect(app.run("ruby -v")).to match("2.5.1")
+      expect(app.output).to match("Ruby version change detected")
     end
   end
 end