+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Master
 
+* Delete the sprockets temp directory for a smaller runtime slug if they are not building assets at runtime (https://github.com/heroku/heroku-buildpack-ruby/pull/812)
+
 ## v195 (10/18/2018)
 
 * Default Ruby version is now 2.4.5 (https://github.com/heroku/heroku-buildpack-ruby/pull/821)